HSEARCH-2581 Clearly separate request processor code from JestClient code
The processor handles the execution order, error handling and
synchonicity, which should be independent from the "transport" layer
provided by JestClient.
Clearly separating the two concepts should allow to replace the
JestClient implementation more easily.
