+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/logging/impl/Log.java
@@ -15,7 +15,7 @@ import org.apache.lucene.search.SortField;
 import org.hibernate.search.analyzer.spi.AnalyzerReference;
 import org.hibernate.search.elasticsearch.cfg.ElasticsearchEnvironment;
 import org.hibernate.search.elasticsearch.client.impl.BackendRequest;
-import org.hibernate.search.elasticsearch.client.impl.BulkRequestFailedException;
+import org.hibernate.search.elasticsearch.processor.impl.BulkRequestFailedException;
 import org.hibernate.search.elasticsearch.schema.impl.ElasticsearchSchemaValidationException;
 import org.hibernate.search.exception.SearchException;
 import org.hibernate.search.util.logging.impl.ClassFormatter;
