+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/processor/impl/BulkRequest.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.client.impl;
+package org.hibernate.search.elasticsearch.processor.impl;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -16,6 +16,8 @@ import java.util.Set;
 
 import org.hibernate.search.backend.IndexingMonitor;
 import org.hibernate.search.backend.LuceneWork;
+import org.hibernate.search.elasticsearch.client.impl.BackendRequest;
+import org.hibernate.search.elasticsearch.client.impl.JestClient;
 import org.hibernate.search.elasticsearch.impl.JestAPIFormatter;
 import org.hibernate.search.elasticsearch.impl.NoOpBackendRequestResultAssessor;
 import org.hibernate.search.elasticsearch.logging.impl.Log;
@@ -34,7 +36,7 @@ import io.searchbox.params.Parameters;
 /**
  * A request group backed by an actual bulk request.
  */
-public class BulkRequest implements ExecutableRequest {
+class BulkRequest implements ExecutableRequest {
 
 	private static final Log LOG = LoggerFactory.make( Log.class );
