+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/processor/impl/BulkRequestFailedException.java
@@ -4,12 +4,13 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.client.impl;
+package org.hibernate.search.elasticsearch.processor.impl;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
+import org.hibernate.search.elasticsearch.client.impl.BackendRequest;
 import org.hibernate.search.exception.SearchException;
 
 import io.searchbox.client.JestResult;
