+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/processor/impl/SingleRequest.java
@@ -4,12 +4,14 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.client.impl;
+package org.hibernate.search.elasticsearch.processor.impl;
 
 import java.util.Collections;
 import java.util.Set;
 
 import org.hibernate.search.backend.IndexingMonitor;
+import org.hibernate.search.elasticsearch.client.impl.BackendRequest;
+import org.hibernate.search.elasticsearch.client.impl.JestClient;
 import org.hibernate.search.exception.ErrorHandler;
 import org.hibernate.search.exception.impl.ErrorContextBuilder;
 
@@ -20,7 +22,7 @@ import io.searchbox.client.JestResult;
  *
  * @author Gunnar Morling
  */
-public class SingleRequest implements ExecutableRequest {
+class SingleRequest implements ExecutableRequest {
 
 	private final JestClient jestClient;
 	private final ErrorHandler errorHandler;
