--- elasticsearch/src/main/resources/META-INF/services/org.hibernate.search.elasticsearch.client.impl.BackendRequestProcessor
@@ -1,1 +1,0 @@
-org.hibernate.search.elasticsearch.client.impl.BackendRequestProcessor
